Line breaks/spaces added, comments removed, return statements simplified. See below for comments on the past few refactor commits.
We changed a couple of things here:
(1) We put all the data elements that we compute into a single class (DataElements), so this contains things like test_corpus, training_labels etc. This way we can give each function the DataElements class as input instead of giving all parameters separately.
(2) We split up the fuctionality of the user-input interaction and the test-data analysis from the classifier functions.
(3) This means that each classifier function (majority_classifier(), rule_based() etc) now just returns the predictions of the sentences (either vectorized or not) that are given to it in the second argument. So for the user-interaction this is a list of size 1, with only the sentence given by the user. And for the test-data analysis, this is the list of sentences from test_corpus.