New structure for storing trained classifiers.
(1) Store the trained classifiers in the DataElement class, to keep all the global data in one place, and to prevent floating variables.
(2) For each call to the classifier, check if it has already been trained, if not, then train it. But in any case, return a trained classifier.
(3) 's' option to manually pretrain the neuralnet removed.