Keywormatching class with levenshtein
With the help of Zan the keyword matcher is finished. It is now able to calculate the levenshtein distance and give the correct word. Also, an additional method is made (keyword_match_info) to handle keyword matching for requests. This should still be implemented in the dialog system. 

Note: changed the name of the class and method it already had to KeywordMatch and keyword_match_pref respectively (also changed it in the transitioner)
Note2: patterns has been removed from the class
Note3: we might still need a fix for assigning values to 'self.preference_values'